requirements.txt
# Core dependencies for telemetry system
pydantic>=2.0.0
python-dateutil>=2.8.0
